+++ R/rnrfa-package.R
@@ -12,20 +12,19 @@
 #' @import rgdal
 #' @import dplyr
 #' @importFrom curl has_internet
-#' @importFrom utils str
-#' @importFrom jsonlite fromJSON
-#' @importFrom lubridate year
-#' @importFrom stats quantile
-#' @importFrom graphics axis legend mtext par plot
-#' @importFrom stats glm
-#' @importFrom httr GET http_error
-#' @importFrom xts xts .indexyear plot.xts
 #' @importFrom ggmap ggmap get_map
 #' @importFrom ggplot2 ggplot geom_point aes coord_flip scale_color_manual theme
 #' @importFrom ggplot2 geom_boxplot theme_minimal ylab xlab unit element_text
 #' @importFrom ggplot2 margin ggtitle
-#' @importFrom sp coordinates proj4string CRS spTransform
+#' @importFrom graphics axis legend mtext par plot
+#' @importFrom httr GET http_error
+#' @importFrom jsonlite fromJSON
+#' @importFrom lubridate year
 #' @importFrom parallel parLapply
+#' @importFrom sp coordinates proj4string CRS spTransform
+#' @importFrom stats aggregate glm quantile
 #' @importFrom tibble as_tibble
+#' @importFrom utils str
+#' @importFrom zoo zoo
 #'
 NULL